+++ test/suite/functional/Generator/Request/GetResourcesRequest.php
@@ -18,6 +18,14 @@ class GetResourcesRequest implements RequestInterface
 
     private ?array $filterByIds = null;
 
+    /** @var string */
+    private $contentType = '';
+
+    public function getContentType(): string
+    {
+        return $this->contentType;
+    }
+
     public function setFilterById(int $filterById): self
     {
         $this->filterById = $filterById;
@@ -40,11 +48,6 @@ class GetResourcesRequest implements RequestInterface
         $this->filterByIds = $filterByIds;
 
         return $this;
-    }
-
-    public function getContentType(): string
-    {
-        return '';
     }
 
     public function getMethod(): string