+++ test/suite/functional/Generator/Request/PatchResourceRequest.php
@@ -14,6 +14,9 @@ class PatchResourceRequest implements RequestInterface
 {
     private PatchResourceRequestBody $patchResourceRequestBody;
 
+    /** @var string */
+    private $contentType = 'application/json';
+
     public function __construct(PatchResourceRequestBody $patchResourceRequestBody)
     {
         $this->patchResourceRequestBody = $patchResourceRequestBody;
@@ -21,7 +24,7 @@ class PatchResourceRequest implements RequestInterface
 
     public function getContentType(): string
     {
-        return 'application/json';
+        return $this->contentType;
     }
 
     public function getMethod(): string
@@ -51,7 +54,7 @@ class PatchResourceRequest implements RequestInterface
 
     public function getHeaders(): array
     {
-        return ['Content-Type' => 'application/json'];
+        return ['Content-Type' => $this->contentType];
     }
 
     /**