+++ test/suite/functional/Generator/Request/PutResourceByIdRequest.php
@@ -13,7 +13,6 @@ use DoclerLabs\ApiClientException\RequestValidationException;
 use Test\Schema\EmbeddedObject;
 use Test\Schema\PutResourceByIdRequestBody;
 use Test\Schema\SerializableInterface;
-use Test\Serializer\ContentType\Json\Json;
 
 class PutResourceByIdRequest implements RequestInterface
 {
@@ -73,13 +72,16 @@ class PutResourceByIdRequest implements RequestInterface
 
     private PutResourceByIdRequestBody $putResourceByIdRequestBody;
 
+    /** @var string */
+    private $contentType = 'application/json';
+
     public function __construct(int $resourceId, int $mandatoryIntegerParameter, string $mandatoryStringParameter, string $mandatoryEnumParameter, DateTimeInterface $mandatoryDateParameter, float $mandatoryFloatParameter, bool $mandatoryBooleanParameter, array $mandatoryArrayParameter, EmbeddedObject $mandatoryObjectParameter, string $xRequestId, PutResourceByIdRequestBody $putResourceByIdRequestBody)
     {
         $this->resourceId                = $resourceId;
         $this->mandatoryIntegerParameter = $mandatoryIntegerParameter;
         $this->mandatoryStringParameter  = $mandatoryStringParameter;
         if (! \in_array($mandatoryEnumParameter, self::ALLOWED_MANDATORY_ENUM_PARAMETER_LIST, true)) {
-            throw new RequestValidationException(\sprintf('Invalid %s value. Given: `%s` Allowed: %s', 'mandatoryEnumParameter', $mandatoryEnumParameter, Json::encode(self::ALLOWED_MANDATORY_ENUM_PARAMETER_LIST)));
+            throw new RequestValidationException(\sprintf('Invalid %s value. Given: `%s` Allowed: %s', 'mandatoryEnumParameter', $mandatoryEnumParameter, \json_encode(self::ALLOWED_MANDATORY_ENUM_PARAMETER_LIST)));
         }
         $this->mandatoryEnumParameter     = $mandatoryEnumParameter;
         $this->mandatoryDateParameter     = $mandatoryDateParameter;
@@ -89,6 +91,11 @@ class PutResourceByIdRequest implements RequestInterface
         $this->mandatoryObjectParameter   = $mandatoryObjectParameter;
         $this->xRequestId                 = $xRequestId;
         $this->putResourceByIdRequestBody = $putResourceByIdRequestBody;
+    }
+
+    public function getContentType(): string
+    {
+        return $this->contentType;
     }
 
     public function setIntegerParameter(int $integerParameter): self
@@ -111,7 +118,7 @@ class PutResourceByIdRequest implements RequestInterface
     public function setEnumParameter(string $enumParameter): self
     {
         if (! \in_array($enumParameter, self::ALLOWED_ENUM_PARAMETER_LIST, true)) {
-            throw new RequestValidationException(\sprintf('Invalid %s value. Given: `%s` Allowed: %s', 'enumParameter', $enumParameter, Json::encode(self::ALLOWED_ENUM_PARAMETER_LIST)));
+            throw new RequestValidationException(\sprintf('Invalid %s value. Given: `%s` Allowed: %s', 'enumParameter', $enumParameter, \json_encode(self::ALLOWED_ENUM_PARAMETER_LIST)));
         }
         $this->enumParameter = $enumParameter;
 
@@ -163,11 +170,6 @@ class PutResourceByIdRequest implements RequestInterface
         return $this;
     }
 
-    public function getContentType(): string
-    {
-        return 'application/json';
-    }
-
     public function getMethod(): string
     {
         return 'PUT';
@@ -203,7 +205,7 @@ class PutResourceByIdRequest implements RequestInterface
 
     public function getHeaders(): array
     {
-        return \array_merge(['Content-Type' => 'application/json'], \array_map(static function ($value) {
+        return \array_merge(['Content-Type' => $this->contentType], \array_map(static function ($value) {
             return $value instanceof SerializableInterface ? $value->toArray() : $value;
         }, \array_filter(['X-Request-ID' => $this->xRequestId], static function ($value) {
             return null !== $value;